refactor: 기본 css 수정 (FE-101)
