+++ src/styles/globals.css
@@ -4,7 +4,7 @@
 
 body {
   /*테스트용 배경이에요*/
-  background-color: #fff;
+  background-color: #1e293b;
 }
 
 button,
